Add a _type helper that checks argument numbers and types.
This is used so that the error messages for using our functions are
_much_ better. The wrapper checks that the arguments passed to the
wrapped function are the right number (so if you forget an argument, you
get an error) and also that they have the right type (so if you pass a
string to a function that expects a number, you get an error).

These are then wrapped around all our public functions.