Distinguish between tAny and tNothing, add explicit return types.
This makes tAny not match undefined, because undefined is not a
meaningful value (and, in particular, if you forget to return from a
function, an undefined is returned, so a function with return type tAny
would match functions that didn't return).

One change that this forced was to add explicit return types to all the
functions - this was intended eventually anyway, but I thought I could
avoid it for now.